requirements.txt
asgiref==3.3.1
Django==3.1.7
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
six~=1.15.0
Pillow~=8.2.0